server/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
python-dotenv
alembic
psycopg2-binary
passlib[bcrypt]
email-validator
pydantic[email]
requests
python-jose
python-multipart
scikit-learn
matplotlib
alpaca-trade-api
apscheduler
tensorflow==2.15.0